requirements.txt
streamlit==1.52.0
PyPDF2==3.0.1
langchain==1.1.0
langchain-core==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
faiss-cpu==1.13.0
google-generativeai==0.8.5